+++ BotDeScans.App/Features/Publish/Pings/1_Everyone.cs
@@ -1,5 +1,4 @@
 ﻿using BotDeScans.App.Features.Publish.Discord;
-using FluentResults;
 using Microsoft.Extensions.Configuration;
 namespace BotDeScans.App.Features.Publish.Pings;
 
@@ -7,6 +6,6 @@ public class EveryonePing(IConfiguration configuration) : Ping(configuration)
 {
     protected override PingType Type => PingType.Everyone;
 
-    public override Task<Result<string>> GetPingAsTextAsync(CancellationToken cancellationToken) =>
-        Task.FromResult(Result.Ok("@everyone"));
+    public override Task<string> GetPingAsTextAsync(CancellationToken cancellationToken) =>
+        Task.FromResult("@everyone");
 }
