+++ BotDeScans.App/Features/Publish/Pings/3_Role.cs
@@ -1,7 +1,7 @@
 ﻿using BotDeScans.App.Extensions;
 using BotDeScans.App.Features.Publish.Discord;
+using BotDeScans.App.Features.Publish.State;
 using BotDeScans.App.Services.Discord;
-using FluentResults;
 using Microsoft.Extensions.Configuration;
 namespace BotDeScans.App.Features.Publish.Pings;
 
@@ -12,11 +12,9 @@ public class RolePing(
 {
     protected override PingType Type => PingType.Role;
 
-    public override async Task<Result<string>> GetPingAsTextAsync(CancellationToken cancellationToken)
+    public override async Task<string> GetPingAsTextAsync(CancellationToken cancellationToken)
     {
         var roleResult = await rolesService.GetRoleFromGuildAsync(publishState.Title.DiscordRoleId.ToString()!, cancellationToken);
-        if (roleResult.IsFailed)
-            return roleResult.ToResult();
 
         return roleResult.Value.ToDiscordString();
     }
