+++ BotDeScans.App/Features/Publish/Pings/4_None.cs
@@ -1,5 +1,4 @@
 ﻿using BotDeScans.App.Features.Publish.Discord;
-using FluentResults;
 using Microsoft.Extensions.Configuration;
 namespace BotDeScans.App.Features.Publish.Pings;
 
@@ -7,6 +6,6 @@ public class NonePing(IConfiguration configuration) : Ping(configuration)
 {
     protected override PingType Type => PingType.None;
 
-    public override Task<Result<string>> GetPingAsTextAsync(CancellationToken cancellationToken) =>
-        Task.FromResult(Result.Ok(string.Empty));
+    public override Task<string> GetPingAsTextAsync(CancellationToken cancellationToken) =>
+        Task.FromResult(string.Empty);
 }
