+++ BotDeScans.App/Features/Publish/PublishService.cs
@@ -1,13 +1,13 @@
 ﻿using BotDeScans.App.Extensions;
+using BotDeScans.App.Features.Publish.State.Models;
 using BotDeScans.App.Features.Publish.Steps;
 using BotDeScans.App.Features.Publish.Steps.Enums;
-using BotDeScans.App.Features.Publish.Steps.Models;
 using Microsoft.Extensions.Configuration;
 namespace BotDeScans.App.Features.Publish;
 
 public class PublishService(IConfiguration configuration, IEnumerable<IStep> steps)
 {
-    public virtual Steps.Models.Steps GetPublishSteps()
+    public virtual EnabledSteps GetEnabledSteps()
     {
         var configurationSteps = configuration
             .GetRequiredValues<StepName>("Settings:Publish:Steps", value => Enum
@@ -15,6 +15,7 @@ public class PublishService(IConfiguration configuration, IEnumerable<IStep> ste
 
         var publishStepsNames = new List<StepName>()
         {
+            StepName.Setup,
             StepName.Download,
             StepName.Compress,
         };
@@ -37,7 +38,7 @@ public class PublishService(IConfiguration configuration, IEnumerable<IStep> ste
                 publishStepsNames.Add(StepName.ZipFiles);
         }
 
-        return new Steps.Models.Steps(steps
+        return new EnabledSteps(steps
             .Where(step => publishStepsNames.Contains(step.Name))
             .OrderBy(step => step.Name)
             .ToDictionary(step => step, step => new StepInfo(step)));
