+++ BotDeScans.App/Features/Publish/Steps/+Dependencies.cs
@@ -17,5 +17,6 @@ internal static class AddDependencies
         .AddScoped<IStep, PdfFilesStep>()
         .AddScoped<IStep, ZipFilesStep>()
         .AddScoped<IStep, CompressFilesStep>()
-        .AddScoped<IStep, DownloadStep>();
+        .AddScoped<IStep, DownloadStep>()
+        .AddScoped<IStep, SetupStep>();
 }
