+++ BotDeScans.App/Features/Publish/Steps/05_UploadZipMegaStep.cs
@@ -1,5 +1,6 @@
 ﻿using BotDeScans.App.Features.Mega;
 using BotDeScans.App.Features.Mega.InternalServices;
+using BotDeScans.App.Features.Publish.State;
 using BotDeScans.App.Features.Publish.Steps.Enums;
 using FluentResults;
 namespace BotDeScans.App.Features.Publish.Steps;
