--- BotDeScans.App/Features/Publish/Steps/Models/StepInfo.cs
@@ -1,17 +1,0 @@
-﻿using BotDeScans.App.Features.Publish.Steps.Enums;
-using FluentResults;
-namespace BotDeScans.App.Features.Publish.Steps.Models;
-
-public class StepInfo(IStep step)
-{
-    public StepStatus StepStatus { get; private set; } = step.Type == StepType.Management 
-         ? StepStatus.QueuedForExecution 
-         : StepStatus.QueuedForValidation;
-
-    public void UpdateStatus(Result result) => 
-        StepStatus = result.IsFailed
-            ? StepStatus.Error
-            : StepStatus == StepStatus.QueuedForValidation
-                ? StepStatus.QueuedForExecution
-                : StepStatus.Success;
-}
