+++ BotDeScans.App/Features/Titles/Create/CreateInteractions.cs
@@ -3,7 +3,9 @@ using BotDeScans.App.Extensions;
 using BotDeScans.App.Infra;
 using BotDeScans.App.Models;
 using BotDeScans.App.Services.Discord;
+using FluentResults;
 using FluentValidation;
+using Remora.Discord.API.Abstractions.Objects;
 using Remora.Discord.Commands.Feedback.Services;
 using Remora.Discord.Interactivity;
 using Remora.Results;
@@ -19,22 +21,22 @@ public class CreateInteractions(
     [Modal(nameof(CreateAsync))]
     [Description("Cadastra nova obra")]
     public async Task<IResult> CreateAsync(
-    string name,
-    string? role)
+        string name,
+        string role)
     {
-        ulong? roleId = null;
-        if (string.IsNullOrEmpty(role) is false)
-        {
-            var roleResult = await rolesService.GetRoleFromGuildAsync(role!, CancellationToken);
-            if (roleResult.IsFailed)
-                return await feedbackService.SendContextualEmbedAsync(
+        // todo: o fluxo vai chamar duas vezes a api do discord aqui para roles,
+        // uma vez para obter o id através de uma string (nome ou id)
+        // outra vez para realizar a validação do objeto Title,
+        // pois tem validação compartilhada com create/edit/publish.
+        // O ideal aqui é criarmos regras do que validar ou não,
+        // E inicialmente podemos, em complemento, criar um cache das roles.
+
+        var roleResult = await GetDiscordRoleId(role);
+        if (roleResult.IsFailed)
+            return await feedbackService.SendContextualEmbedAsync(
                 embed: EmbedBuilder.CreateErrorEmbed(roleResult.ToResult()),
                 ct: CancellationToken);
-
-            roleId = roleResult.Value.ID.Value;
-        }
-
-        var title = new Title(name, roleId);
+        var title = new Title(name, roleResult.Value?.ID.Value);
         var validationResult = await validator.ValidateAsync(title);
         if (validationResult.IsValid is false)
             return await feedbackService.SendContextualEmbedAsync(
@@ -47,5 +49,17 @@ public class CreateInteractions(
         return await feedbackService.SendContextualEmbedAsync(
             embed: EmbedBuilder.CreateSuccessEmbed("Obra cadastrada com sucesso!"),
             ct: CancellationToken);
+    }
+
+    private async Task<FluentResults.Result<IRole?>> GetDiscordRoleId(string roleRequest)
+    {
+        if (string.IsNullOrWhiteSpace(roleRequest))
+            return FluentResults.Result.Ok<IRole?>(null);
+
+        var roleResult = await rolesService.GetRoleFromGuildAsync(roleRequest, CancellationToken);
+        if (roleResult.IsFailed)
+            return roleResult.ToResult();
+
+        return FluentResults.Result.Ok<IRole?>(roleResult.Value);
     }
 }
