+++ BotDeScans.App/Services/GoogleBloggerService.cs
@@ -1,5 +1,5 @@
 ﻿using BotDeScans.App.Extensions;
-using BotDeScans.App.Features.Publish;
+using BotDeScans.App.Features.Publish.State;
 using Google.Apis.Blogger.v3;
 using Google.Apis.Blogger.v3.Data;
 using Microsoft.Extensions.Configuration;
